Specify file separation
