+++ .gitignore
@@ -1,0 +1,5 @@
+# Ignore the virtual environment directory
+.venv/
+
+# Ignore the Git metadata directory
+.git/
