+++ scripts/preprocess.py/CameraCalibrateImage.py
