+++ test/Motor/camera_test.py
