+++ test/OpenCV/test_checkboard.py
@@ -1,0 +1,48 @@
+import cv2
+import numpy as np
+
+# 체스보드 패턴 크기
+chessboard_size = (9, 6)
+
+# 체스보드찾기 위한 준비
+criteria = (cv2.TermCriteria_EPS + cv2.TermCriteria_MAX_ITER, 30, 0.001)
+objp = np.zeros((chessboard_size[0] * chessboard_size[1], 3), np.float32)
+objp[:, :2] = np.mgrid[0:chessboard_size[0], 0:chessboard_size[1]].T.reshape(-1, 2)
+
+# 실제 3D 점과 이미지의 2D 점 저장 리스트
+objpoints = []  # 3D 점
+imgpoints = []  # 2D 점
+
+# 카메라 캘리브레이션을 위해 여러 이미지를 로드
+images = ["image1.jpg", "image2.jpg", "image3.jpg"]
+for fname in images:
+    img = cv2.imread(fname)
+    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+    
+    # 체스보드 코너 찾기
+    ret, corners = cv2.findChessboardCorners(gray, chessboard_size, None)
+    if ret:
+        objpoints.append(objp)
+        corners2 = cv2.cornerSubPix(gray, corners, (11, 11), (-1, -1), criteria)
+        imgpoints.append(corners2)
+        
+        # 코너 그리기
+        cv2.drawChessboardCorners(img, chessboard_size, corners2, ret)
+        cv2.imshow('Corners', img)
+        cv2.waitKey(500)
+
+cv2.destroyAllWindows()
+
+# 카메라 보정
+ret, mtx, dist, rvecs, tvecs = cv2.calibrateCamera(objpoints, imgpoints, gray.shape[::-1], None, None)
+
+# 보정된 이미지 출력
+img = cv2.imread('test_image.jpg')
+h, w = img.shape[:2]
+newcameramtx, roi = cv2.getOptimalNewCameraMatrix(mtx, dist, (w, h), 1, (w, h))
+dst = cv2.undistort(img, mtx, dist, None, newcameramtx)
+
+# 결과 확인
+cv2.imshow('Undistorted Image', dst)
+cv2.waitKey(0)
+cv2.destroyAllWindows()
